Add separate URL for client to push symbol packages.
.NET 5 packages seem to include source and symbols in normal packages.
This confuses Klondike since it can't tell how to treat them.

Currently, any package that has src and lib/**/*.pdb will be redirected
to push packages to /api/symbols.

Later, this redirect may be removed and these packages will be treated
the same as packages that don't have source and symbols.